+++ src/core/modules.py
@@ -142,7 +142,7 @@ func read
 python("self.io_read()")
 endfunc
 func py_load_file($path)
-python("import importlib.util\; spec = importlib.util.spec_from_file_location('pyloadedfile', self.get_var('path'))\; m = importlib.util.module_from_spec(spec)\; spec.loader.exec_module(m)\; self.mem = m")
+python("import importlib.util; spec = importlib.util.spec_from_file_location('pyloadedfile', self.get_var('path')); m = importlib.util.module_from_spec(spec); spec.loader.exec_module(m); self.mem = m")
 endfunc
 func fopen($args)
 if typeof($args) != tuple